+++ src/App.vue
@@ -64,16 +64,16 @@ export default {
         json: [
             {ch: { value: '1', checked: true }, date: '12.08.1999', user: 'admin1', orderno: '1234', href: 'dasd'},
             {ch: '2', date: '13.08.1999', user: 'use2', orderno: '3123'},
-            {ch: '3', date: '14.08.1999', user: 'admin3', orderno: '415'},
+            {ch: '3', date: '14.08.2002', user: 'admin3', orderno: '415'},
             {ch: '4', date: '12.08.1999', user: 'admin4', orderno: '1234', href: 'dasd'},
             {ch: '5', date: '13.08.1999', user: 'user', orderno: '3123'},
-            {ch: '6', date: '14.08.1999', user: 'admin', orderno: '415'},
+            {ch: '6', date: '14.08.1970', user: 'admin', orderno: '415'},
             {ch: '7', date: '15.08.1999', user: 'user', orderno: '412412'},
-            {ch: '8', date: '12.08.1999', user: 'admin', orderno: '1234', href: 'dasd'},
+            {ch: '8', date: '12.08.2000', user: 'admin', orderno: '1234', href: 'dasd'},
             {ch: '9', date: '13.08.1999', user: 'user', orderno: '3123'},
-            {ch: '10', date: '14.08.1999', user: 'admin', orderno: '415'},
+            {ch: '10', date: '14.08.1998', user: 'admin', orderno: '415'},
             {ch: '11', date: '15.08.1999', user: 'user', orderno: '412412'},
-            {ch: '12', date: '15.08.1999', user: 'user', orderno: '412412'},
+            {ch: '12', date: '15.08.2015', user: 'user', orderno: '412412'},
         ]
     }
   },
